move utilities.js to utilities folder
